chatbot/requirements.txt
requests
beautifulsoup4
lxml
pypdf
ddgs
pandas
numpy
scikit-learn
nltk
unidecode
langdetect
joblib
tqdm
sentence-transformers
faiss-cpu
transformers
torch
python-dotenv